requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
langchain==0.0.335
openai==1.2.4
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.1
scikit-learn==1.3.2
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.18.0
nltk==3.8.1
spacy==3.7.2
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
